Clasificador-de-digitos-propios/requirements.txt
streamlit
numpy
tensorflow
Pillow
matplotlib
pandas